norm fixes from left to right. Made more files for the expand functions and added all to the minishell.h header file
