+++ src/ft_env.c
@@ -6,7 +6,7 @@
 /*   By: plang <plang@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/23 13:54:24 by plang             #+#    #+#             */
-/*   Updated: 2024/05/09 15:35:22 by dbarrene         ###   ########.fr       */
+/*   Updated: 2024/06/05 17:25:59 by plang            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,8 +25,8 @@ int	ft_env(t_env **env)
 				printf("%s\n", temp->env_element);
 			temp = temp->next;
 		}
-		return (0); //EXIT_SUCCESS
+		return (0);
 	}
 	else
-		return (1); //EXIT_FAILURE
+		return (1);
 }
