+++ src/ft_unset.c
@@ -6,7 +6,7 @@
 /*   By: plang <plang@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/24 10:37:22 by plang             #+#    #+#             */
-/*   Updated: 2024/05/10 13:14:00 by plang            ###   ########.fr       */
+/*   Updated: 2024/06/05 17:26:58 by plang            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -108,5 +108,5 @@ int	ft_unset(t_env **env, char **cm_ar)
 		}
 		i++;
 	}
-	return (0);//EXIT_SUCCESS
+	return (0);
 }
