+++ src/valid_redirect.c
@@ -6,25 +6,11 @@
 /*   By: plang <plang@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 14:47:32 by plang             #+#    #+#             */
-/*   Updated: 2024/05/24 17:00:12 by plang            ###   ########.fr       */
+/*   Updated: 2024/06/05 17:15:40 by plang            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
-
-// void	skipping_through_quotes(char *str, int *i)
-// {
-// 	char	temp;
-
-// 	temp = 0;
-// 	if (str[*i] == '"' || str[*i] == '\'')
-// 	{
-// 		temp = str[*i];
-// 		*i += 1;
-// 		while (str[*i] != temp && str[*i] != '\0')
-// 			*i += 1;
-// 	}
-// }
 
 int	we_have_redirect(char *str, int *i)
 {
